implement taskService and add mapToTask method in TaskMapper. Add a method in TaskRepository for find tasks by their priority

+++ tasks_manager/src/main/java/com/example/demo/dto/response/ProjectResponseDTO.java
@@ -7,7 +7,7 @@ public record ProjectResponseDTO(
 		String name, 
 		String description, 
 		LocalDate creationDate,
-		Long taskEseguiti,
-		Long taskInEsecuzione,
-		Long taskDaEseguire
+		int taskEseguiti,
+		int taskInEsecuzione,
+		int taskDaEseguire
 		) {}
